[Feature] Added per-call headers capabilities to OpenAiClient (#25)
# Description
In order to better facilitate integration with OpenAI proxies, we want
to be able to inject headers per request so that these headers could be
used for auxiliary handling of actual OpenAI calls e.g use case being
usage tracking and rate limits.

# Approach
The approach is to retain much of the OpenAiClient interface and calls
but extending the capabilities to send extra headers per request. This
requires us to open up the `abstract` methods to take in a new inner
class in `OpenAiClient.OpenAiClientContext` as a method argument. We can
extend this context so that other implementations of `OpenAiClient`
abstract class could use the same context object to add in other
metadata if needed for executing OpenAi API methods.

For now, I've only added `headers` in the `OpenAiClientContext` but I
could imagine this being used for other cases as well. The default
signature is retained for the `OpenAiClient` so we can still call
`chatCompletion(ChatCompletionRequest)` as it is and it will simply pass
along a default `OpenAiClientContext`

cc: @langchain4j